# frontend-beer-frontier/openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
  /classifications/{id}:
    get:
      tags:
        - classification-controller
      operationId: getClassificationById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Classification'
    put:
      tags:
        - classification-controller
      operationId: updateClassification
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Classification'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Classification'
  /breweries/{id}:
    get:
      tags:
        - brewery-controller
      operationId: getBreweryById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Brewery'
    put:
      tags:
        - brewery-controller
      operationId: updateBrewery
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Brewery'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Brewery'
  /beers/{id}:
    get:
      tags:
        - beer-controller
      operationId: getBeer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Beer'
    put:
      tags:
        - beer-controller
      operationId: updateBeer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Beer'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Beer'
    delete:
      tags:
        - beer-controller
      operationId: deleteBeer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
  /beers:
    get:
      tags:
        - beer-controller
      operationId: getBeers
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Beer'
    post:
      tags:
        - beer-controller
      operationId: addBeer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Beer'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Beer'
  /user/info:
    get:
      tags:
        - user-controller
      operationId: userInfo
      responses:
        '200':
          description: OK
          content: {}
  /classifications:
    get:
      tags:
        - classification-controller
      operationId: getClassification
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Classification'
  /breweries:
    get:
      tags:
        - brewery-controller
      operationId: getBreweries
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Brewery'
  /beers/by-brewery/{breweryId}:
    get:
      tags:
        - beer-controller
      operationId: getBeersByBreweryId
      parameters:
        - name: breweryId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Beer'
  /breweries/{breweryId}/beers/{beerId}:
    delete:
      tags:
        - brewery-controller
      operationId: unlinkBeer
      parameters:
        - name: breweryId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: beerId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
components:
  schemas:
    Beer:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        alcoholPercentage:
          type: integer
          format: int32
        beerType:
          type: string
          enum:
            - ALCOHOL_FREE
            - ALE
            - FRUIT
            - CHERRY
        brewery:
          $ref: '#/components/schemas/Brewery'
        classification:
          $ref: '#/components/schemas/Classification'
    Brewery:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        location:
          type: string
    Classification:
      type: object
      properties:
        id:
          type: integer
          format: int64
        country:
          type: string
          enum:
            - BELGIUM
            - GERMANY
            - AMERICA
            - SPAIN
            - FRANCE
        usedGrainType:
          type: string
          enum:
            - BARELY
            - WHEAT
            - CORN
            - BASE_MALT
            - OATS
        fermentationType:
          type: string
          enum:
            - WARM
            - COOL
            - WILD
            - SPONTANEOUS
        namesAndOrigins:
          type: string
          enum:
            - ABBEY_BEER
            - TRAPPIST_BEER
            - DARK_BEER
        beers:
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/Beer'