# frontend-beer-frontier/src/app/api/models/Brewery.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface Brewery
 */
export interface Brewery {
    /**
     * 
     * @type {number}
     * @memberof Brewery
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof Brewery
     */
    name?: string;
    /**
     * 
     * @type {string}
     * @memberof Brewery
     */
    location?: string;
}

/**
 * Check if a given object implements the Brewery interface.
 */
export function instanceOfBrewery(value: object): value is Brewery {
    return true;
}

export function BreweryFromJSON(json: any): Brewery {
    return BreweryFromJSONTyped(json, false);
}

export function BreweryFromJSONTyped(json: any, ignoreDiscriminator: boolean): Brewery {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'name': json['name'] == null ? undefined : json['name'],
        'location': json['location'] == null ? undefined : json['location'],
    };
}

export function BreweryToJSON(value?: Brewery | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'id': value['id'],
        'name': value['name'],
        'location': value['location'],
    };
}